+++ core/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/submodeltemplate/SubmodelTemplateProcessorTest.java
@@ -16,28 +16,35 @@ package de.fraunhofer.iosb.ilt.faaast.service.submodeltemplate;
 
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import de.fraunhofer.iosb.ilt.faaast.service.Service;
-import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.AssetConnectionException;
 import de.fraunhofer.iosb.ilt.faaast.service.config.CoreConfig;
-import de.fraunhofer.iosb.ilt.faaast.service.exception.ConfigurationException;
 import de.fraunhofer.iosb.ilt.faaast.service.exception.MessageBusException;
 import de.fraunhofer.iosb.ilt.faaast.service.filestorage.FileStorage;
 import de.fraunhofer.iosb.ilt.faaast.service.messagebus.MessageBus;
+import de.fraunhofer.iosb.ilt.faaast.service.model.AASFull;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.paging.Page;
-import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementCreateEventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementDeleteEventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementUpdateEventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.persistence.Persistence;
+import de.fraunhofer.iosb.ilt.faaast.service.util.ReferenceBuilder;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import org.eclipse.digitaltwin.aas4j.v3.model.Environment;
 import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.ArgumentMatcher;
 import org.mockito.invocation.InvocationOnMock;
 
 
@@ -52,94 +59,88 @@ public class SubmodelTemplateProcessorTest {
 
     @Before
     public void init() throws MessageBusException {
-        /*
-         * processor = mock(SubmodelTemplateProcessor.class);
-         * mockMessageBus();
-         * environment = AASFull.createEnvironment();
-         * submodel0 = environment.getSubmodels().get(0);
-         * submodel1 = environment.getSubmodels().get(1);
-         * ArgumentMatcher<Submodel> isSubmodel0 = submodel -> Objects.equals(submodel, submodel0);
-         * when(processor.accept(argThat(isSubmodel0))).thenReturn(true);
-         * when(processor.add(eq(submodel0), any())).thenReturn(true);
-         * when(processor.update(eq(submodel0), any())).thenReturn(true);
-         * when(processor.delete(eq(submodel0), any())).thenReturn(true);
-         */
+        processor = mock(SubmodelTemplateProcessor.class);
+        mockMessageBus();
+        environment = AASFull.createEnvironment();
+        submodel0 = environment.getSubmodels().get(0);
+        submodel1 = environment.getSubmodels().get(1);
+        ArgumentMatcher<Submodel> isSubmodel0 = submodel -> Objects.equals(submodel, submodel0);
+        when(processor.accept(argThat(isSubmodel0))).thenReturn(true);
+        when(processor.add(eq(submodel0), any())).thenReturn(true);
+        when(processor.update(eq(submodel0), any())).thenReturn(true);
+        when(processor.delete(eq(submodel0), any())).thenReturn(true);
     }
 
 
     @Test
-    public void testStart() throws ConfigurationException, AssetConnectionException, PersistenceException, MessageBusException {
-        /*
-         * List<Submodel> submodels = environment.getSubmodels();
-         * createService(submodels);
-         * verify(processor, times(submodels.size())).accept(any());
-         * verify(processor, times(1)).add(eq(submodel0), any());
-         * verify(processor, times(0)).add(eq(submodel1), any());
-         * verify(processor, times(0)).update(any(), any());
-         * Assert.assertNotNull(service);
-         */
+    public void testStart() throws Exception {
+        List<Submodel> submodels = environment.getSubmodels();
+        createService(submodels);
+        verify(processor, times(submodels.size())).accept(any());
+        verify(processor, times(1)).add(eq(submodel0), any());
+        verify(processor, times(0)).add(eq(submodel1), any());
+        verify(processor, times(0)).update(any(), any());
+        Assert.assertNotNull(service);
     }
 
 
     @Test
-    public void testUpdate() throws ConfigurationException, AssetConnectionException, PersistenceException, MessageBusException {
-        /*
-         * List<Submodel> submodels = new ArrayList<>();
-         * createService(submodels);
-         * // Send update event to MessageBus
-         * ElementUpdateEventMessage msg = new ElementUpdateEventMessage();
-         * msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
-         * msg.setValue(submodel0);
-         * service.getMessageBus().publish(msg);
-         * // called for every Submodel in createService and for the Submodel to update
-         * verify(processor, times(submodels.size() + 1)).accept(any());
-         * verify(processor, times(1)).update(eq(submodel0), any());
-         * Assert.assertNotNull(service);
-         */
+    public void testUpdate() throws Exception {
+        List<Submodel> submodels = new ArrayList<>();
+        createService(submodels);
+
+        // Send update event to MessageBus
+        ElementUpdateEventMessage msg = new ElementUpdateEventMessage();
+        msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
+        msg.setValue(submodel0);
+        service.getMessageBus().publish(msg);
+        // called for every Submodel in createService and for the Submodel to update
+        verify(processor, times(submodels.size() + 1)).accept(any());
+        verify(processor, times(1)).update(eq(submodel0), any());
+        Assert.assertNotNull(service);
     }
 
 
     @Test
-    public void testDelete() throws ConfigurationException, AssetConnectionException, PersistenceException, MessageBusException {
-        /*
-         * List<Submodel> submodels = new ArrayList<>();
-         * createService(submodels);
-         * // Send delete event to MessageBus
-         * ElementDeleteEventMessage msg = new ElementDeleteEventMessage();
-         * msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
-         * msg.setValue(submodel0);
-         * service.getMessageBus().publish(msg);
-         * // called for every Submodel in createService and for the Submodel to delete
-         * verify(processor, times(submodels.size() + 1)).accept(any());
-         * verify(processor, times(1)).delete(eq(submodel0), any());
-         * Assert.assertNotNull(service);
-         */
+    public void testDelete() throws Exception {
+        List<Submodel> submodels = new ArrayList<>();
+        createService(submodels);
+
+        // Send delete event to MessageBus
+        ElementDeleteEventMessage msg = new ElementDeleteEventMessage();
+        msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
+        msg.setValue(submodel0);
+        service.getMessageBus().publish(msg);
+        // called for every Submodel in createService and for the Submodel to delete
+        verify(processor, times(submodels.size() + 1)).accept(any());
+        verify(processor, times(1)).delete(eq(submodel0), any());
+        Assert.assertNotNull(service);
     }
 
 
     @Test
-    public void testCreate() throws ConfigurationException, AssetConnectionException, PersistenceException, MessageBusException {
-        /*
-         * List<Submodel> submodels = new ArrayList<>();
-         * createService(submodels);
-         * // Send create event to MessageBus
-         * ElementCreateEventMessage msg = new ElementCreateEventMessage();
-         * msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
-         * msg.setValue(submodel0);
-         * service.getMessageBus().publish(msg);
-         * // called for every Submodel in createService and for the Submodel to delete
-         * verify(processor, times(submodels.size() + 1)).accept(any());
-         * verify(processor, times(1)).add(eq(submodel0), any());
-         * Assert.assertNotNull(service);
-         */
+    public void testCreate() throws Exception {
+        List<Submodel> submodels = new ArrayList<>();
+        createService(submodels);
+
+        // Send create event to MessageBus
+        ElementCreateEventMessage msg = new ElementCreateEventMessage();
+        msg.setElement(ReferenceBuilder.forSubmodel(submodel0));
+        msg.setValue(submodel0);
+        service.getMessageBus().publish(msg);
+        // called for every Submodel in createService and for the Submodel to delete
+        verify(processor, times(submodels.size() + 1)).accept(any());
+        verify(processor, times(1)).add(eq(submodel0), any());
+        Assert.assertNotNull(service);
     }
 
 
-    private void createService(List<Submodel> submodels) throws AssetConnectionException, MessageBusException, ConfigurationException, PersistenceException {
+    private void createService(List<Submodel> submodels) throws Exception {
         Persistence persistence = mock(Persistence.class);
         FileStorage fileStorage = mock(FileStorage.class);
         when(persistence.getAllSubmodels(any(), any())).thenReturn(Page.of(submodels));
         service = new Service(CoreConfig.DEFAULT, persistence, fileStorage, messageBus, List.of(), List.of(), List.of(processor));
+        service.start();
     }
 
 
@@ -149,7 +150,7 @@ public class SubmodelTemplateProcessorTest {
         doAnswer((InvocationOnMock invocation) -> {
             ElementCreateEventMessage eventMessage = invocation.getArgument(0);
             try {
-                service.handleCreateEvent(eventMessage);
+                service.getSubmodelTemplateManager().handleCreateEvent(eventMessage);
             }
             catch (Exception e) {
                 fail();
@@ -159,7 +160,7 @@ public class SubmodelTemplateProcessorTest {
         doAnswer((InvocationOnMock invocation) -> {
             ElementUpdateEventMessage eventMessage = invocation.getArgument(0);
             try {
-                service.handleUpdateEvent(eventMessage);
+                service.getSubmodelTemplateManager().handleUpdateEvent(eventMessage);
             }
             catch (Exception e) {
                 fail();
@@ -170,7 +171,7 @@ public class SubmodelTemplateProcessorTest {
         doAnswer((InvocationOnMock invocation) -> {
             ElementDeleteEventMessage eventMessage = invocation.getArgument(0);
             try {
-                service.handleDeleteEvent(eventMessage);
+                service.getSubmodelTemplateManager().handleDeleteEvent(eventMessage);
             }
             catch (Exception e) {
                 fail();