+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/ServiceSpecificationProfile.java
@@ -463,11 +463,11 @@ public enum ServiceSpecificationProfile {
                     GenerateSerializationByIdsRequest.class,
                     GetSelfDescriptionRequest.class)),
     FAAAST_IMPORT(
-            "https://github.com/FraunhoferIOSB/FAAAST-Service/API/1/3/Import",
+            "https://github.com/FraunhoferIOSB/FAAAST-Service/API/1/4/Import",
             List.of(),
             List.of(ImportRequest.class)),
     FAAAST_RESET(
-            "https://github.com/FraunhoferIOSB/FAAAST-Service/API/1/3/Reset",
+            "https://github.com/FraunhoferIOSB/FAAAST-Service/API/1/4/Reset",
             List.of(),
             List.of(ResetRequest.class));