+++ docs/source/other/release-notes.md
@@ -1,12 +1,24 @@
 # Release Notes
 <!--start:changelog-header-->
-## 1.3.0-SNAPSHOT (current development version)<!--end:changelog-header-->
+## 1.4.0-SNAPSHOT (current development version)<!--end:changelog-header-->
+## 1.3.0
+
+**New Features & Major Changes**
+- General
+	- New interface `SubmodelTemplateProcessor` that enables handling of SMTs
+- SubmodelTemplateProcessors
+	- Added processor for SMT Asset Interfaces Description (AID) and Asset Interfaces Mapping Configuration (AIMC), allowing to create/update/delete asset connections on-the-fly
+
 
 **Internal changes & bugfixes**
 - General
 	- Fixed bug that auxiliary files were not loaded when starting from code with an initial model file
 	- Fixed bug that caused deleting submodel-refs from AAS to fail when the submodel-ref had referredSemanticId set
 	- Minor corrections in Logging
+	- Make minInflateRatio configurable to be able to prevent zip bomb error when loading AASX files.
+- Asset Connection
+	- Asset connections are now normalized, i.e., connections with exactly the same properties (e.g. type, server, credentials, etc.) are merged into a single connection
+	- Direct modification of asset connections and providers no longer supported, instead every interaction with connections or providers must be made via `AssetConnectionManager`
 - Endpoint
 	- HTTP
 		- URL query parameters are now correctly URL-decoded